Near zero memory construction

+++ src/construct/indexedleaf.cpp
@@ -3,7 +3,8 @@
 
 struct IndexedLeaf {
     const Leaf leaf;
-    const unsigned int index;
-    IndexedLeaf(const Leaf leaf, const unsigned int index) : leaf(leaf), index(index) {}
+    const unsigned long long int index;
+    IndexedLeaf(const Leaf leaf, const unsigned long long int index) 
+            : leaf(leaf), index(index) {}
 };
